Implemented utility class for camera lens
